@@ -10,22 +10,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.0",
-    "@angular/common": "^14.1.0",
-    "@angular/compiler": "^14.1.0",
-    "@angular/core": "^14.1.0",
-    "@angular/forms": "^14.1.0",
-    "@angular/platform-browser": "^14.1.0",
-    "@angular/platform-browser-dynamic": "^14.1.0",
-    "@angular/router": "^14.1.0",
+    "@angular/animations": "^15.2.5",
+    "@angular/common": "^15.2.5",
+    "@angular/compiler": "^15.2.5",
+    "@angular/core": "^15.2.5",
+    "@angular/forms": "^15.2.5",
+    "@angular/localize": "^15.2.5",
+    "@angular/platform-browser": "^15.2.5",
+    "@angular/platform-browser-dynamic": "^15.2.5",
+    "@angular/router": "^15.2.5",
+    "@ng-bootstrap/ng-bootstrap": "^14.1.0",
+    "@popperjs/core": "^2.11.7",
+    "bootstrap": "^5.2.3",
+    "popper.js": "^1.16.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.1.3",
-    "@angular/cli": "~14.1.3",
-    "@angular/compiler-cli": "^14.1.0",
+    "@angular-devkit/build-angular": "^15.2.5",
+    "@angular/cli": "~15.2.5",
+    "@angular/compiler-cli": "^15.2.5",
     "@types/jasmine": "~4.0.0",
     "jasmine-core": "~4.2.0",
     "karma": "~6.4.0",
@@ -33,6 +38,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.7.2"
+    "typescript": "~4.9.5"
   }
 }
