@@ -3,6 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
+import { NgbPaginationModule, NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -10,20 +11,32 @@ import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { routing } from './app.routing';
 import { SidebarComponent } from './components/sidebar/sidebar.component';
 import { FooterComponent } from './components/footer/footer.component';
+import { LoginComponent } from './components/login/login.component';
+import { TopComponent } from './components/top/top.component';
+import { IndexColaboradorComponent } from './components/colaborador/index-colaborador/index-colaborador.component';
+import { CreateColaboradorComponent } from './components/colaborador/create-colaborador/create-colaborador.component';
+import { EditColaboradorComponent } from './components/colaborador/edit-colaborador/edit-colaborador.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     DashboardComponent,
     SidebarComponent,
-    FooterComponent
+    FooterComponent,
+    LoginComponent,
+    TopComponent,
+    IndexColaboradorComponent,
+    CreateColaboradorComponent,
+    EditColaboradorComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     routing,
     FormsModule,
-    HttpClientModule
+    HttpClientModule,
+    NgbPaginationModule,
+    NgbModule
   ],
   providers: [],
   bootstrap: [AppComponent]
